editor stuff

+++ src/main/java/sep/fimball/view/pinballcanvas/PinballCanvasSubView.java
@@ -88,7 +88,7 @@ public class PinballCanvasSubView implements ViewBoundToViewModel<PinballCanvasV
         graphicsContext.fillRect(0, 0, canvas.getWidth(), canvas.getHeight());
 
         graphicsContext.save();
-        graphicsContext.translate(canvas.getWidth() / 2d - cameraPosition.get().getX() * Config.pixelsPerGridUnit * cameraZoom.get(), canvas.getHeight() / 2d - cameraPosition.get().getY() * Config.pixelsPerGridUnit * cameraZoom.get());
+        graphicsContext.translate((int) (canvas.getWidth() / 2d - cameraPosition.get().getX() * Config.pixelsPerGridUnit * cameraZoom.get()), (int) (canvas.getHeight() / 2d - cameraPosition.get().getY() * Config.pixelsPerGridUnit * cameraZoom.get()));
 
         graphicsContext.scale(cameraZoom.get(), cameraZoom.get());
